Fix card spacing in information section
This commit fixes the card spacing in the information section.
Before the outer edges of the contents in the information section were
not in line horizontally with the outer edges of the content in the
quote section of the webpage.
Now they are which is the desired outcome.

This commit also removes some unnecessary comments in the css stylesheet
of the webpage.